+++ src/main/java/act/util/CsvView.java
@@ -1,4 +1,4 @@
-package act.cli;
+package act.util;
 
 /*-
  * #%L
@@ -20,7 +20,7 @@ package act.cli;
  * #L%
  */
 
-import act.util.PropertySpec;
+import act.cli.TableView;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -28,31 +28,26 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * **Note** this is deprecated, please use `act.util.CsvView` instead
- *
  * Mark a command method return value shall be displayed using a csv file format. e.g
- *
- * ```
+ * <pre>
  * id,onetime,trigger
  * __act_app_stop,true,null
  * __act_app_app_act_plugin_loaded,true,null
- * ```
- *
- *
- * {@code CsvView} can be used in conjunction with {@link PropertySpec}
- * to export only specified fields
- *
- *
- * Note if a method is marked with neither {@link CsvView} nor
- * {@link JsonView} then the console will simply use
- * {@link Object#toString()} to present the data.
- *
+ * </pre>
+ * <p>
+ *     {@code CsvView} can be used in conjunction with {@link PropertySpec}
+ *     to export only specified fields
+ * </p>
+ * <p>
+ *     Note if a method is marked with neither {@link CsvView} nor
+ *     {@link JsonView} then the console will simply use
+ *     {@link Object#toString()} to present the data.
+ * </p>
  * @see JsonView
  * @see TableView
  * @see PropertySpec
  */
 @Retention(RetentionPolicy.CLASS)
 @Target(ElementType.METHOD)
-@Deprecated
 public @interface CsvView {
 }